fix(ping): resolve 12 Kotlin compile errors from L0-L3 architecture
- BenchmarkEngine: pass profile.id (String) to recordTransportVerdict
- MarbleIntelligence: move carryLayerColumns outside ContentValues.apply
  (values not visible inside its own initializer); explicit Any? arrayOf for
  execSQL bind args; explicit ArrayList type in getOrPut
- MultiVectorReachability: replace repeat{...break} with for loop;
  fix Int/Long mix in epoch rotation seed
- ProtocolFingerprintAwareVerifier: qualify SAWTOOTH_HIGH_CONFIDENCE from
  SawtoothDetector (separate object)
- SocksHttpClient: cast port bytes to Byte; import SSLException
- Aether2026: Aether.Pearl -> Aether.Ink (no Pearl token)
- MarbleDesignSystem: hoist Aether.Danger before non-composable DrawScope
- MarbleVpnService: helper resolves repo via application cast; import
  ProtocolFingerprintAwareVerifier

+++ app/src/main/java/com/marbleng/app/core/SocksHttpClient.kt
@@ -10,6 +10,7 @@ import java.net.Socket
 import java.net.URL
 import java.util.Locale
 import java.util.zip.GZIPInputStream
+import javax.net.ssl.SSLException
 import javax.net.ssl.SSLSocket
 import javax.net.ssl.SSLSocketFactory
 import kotlin.math.min
@@ -680,7 +681,7 @@ object SocksHttpClient {
             val target = socksTarget(host)
             output.write(byteArrayOf(5, 1, 0, target.first.toByte()))
             output.write(target.second)
-            output.write(byteArrayOf(443 ushr 8, 443 and 0xff))
+            output.write(byteArrayOf((443 ushr 8).toByte(), (443 and 0xff).toByte()))
             output.flush()
 
             val reply = ByteArray(4)
